Reverse a couple changes.
There is a contradiction between what make check-plist wants and what make install needs.  If I do not tick server in the config but the pkg-plist does not have %%SERVER%% tags on etc/rc.d/luanti then the install fails because it cannot find the file staged. However, if I run the check-plist then it insists that etc/rc.d/luanti should be present in the pkg-plist.  It cannot be both, so I am including the tags and ignoring that check-plist prefers there be no tags.